requirements.txt
absl-py==2.0.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
colorama==0.4.6
docopt==0.6.2
flask
google-ai-generativelanguage==0.4.0
google-api-core==2.24.0
google-auth==2.37.0
google-generativeai==0.3.2
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.62.3
idna==3.10
Js2Py==0.74
packaging==24.2
pipwin==0.5.2
proto-plus==1.25.0
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pydub==0.25.1
pyjsparser==2.7.1
PyPDF2==3.0.1
PyPrind==2.11.3
pySmartDL==1.3.4
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
six==1.17.0
soupsieve==2.6
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
websockets==14.1
PyQt6>=6.4.0
exceptiongroup==1.2.0
